separate out generate case tests (doh!)
